separate veteran_service_officer and attorney (other) types into their own controllers
