+++ modules/veteran/config/routes.rb
@@ -4,7 +4,8 @@ Veteran::Engine.routes.draw do
   match '/v0/*path', to: 'application#cors_preflight', via: [:options]
   namespace :v0, defaults: { format: 'json' } do
     get 'representatives/find_rep', to: 'representatives#find_rep'
-    resources :accredited_representatives, only: %i[index]
+    resources :vso_accredited_representatives, only: %i[index]
+    resources :other_accredited_representatives, only: %i[index]
     get 'apidocs', to: 'apidocs#index'
   end
 end
